Backend/requirements.txt
beautifulsoup4
certifi
charset-normalizer
click
colorama
filelock
Flask
huggingface-hub
idna
itsdangerous
Jinja2
lxml
MarkupSafe
newspaper3k              # keep only if scraping is still needed
nltk                     # if actually used (I didn’t see it in code)
packaging
regex
requests
safetensors
setuptools
six
tokenizers
torch
tqdm
transformers
urllib3
Werkzeug
flask-caching
flask-cors
gunicorn
python-dotenv
